feat(1078): amend file structure
