+++ packages/common/src/view-model-maps/rows/def.d.ts
@@ -3,7 +3,7 @@ import type { Api } from 'appeals-service-api';
 export interface Row {
 	keyText: string;
 	valueText: string;
-	condition: (caseData: Api.AppealCaseWithAppellant) => boolean | undefined;
+	condition: (caseData: Api.AppealCaseWithAppellant) => string | boolean | undefined;
 }
 
 export type Rows = Array<Row>;
